feat: minor fix in checkout
account for more than one word in city value
