+++ src/components/checkout/PaymentForm.tsx
@@ -1,9 +1,8 @@
 import { ReactNode } from 'react'
 import { SummaryCard } from './SummaryCard'
-import { CartItemProps, StoreProps } from '@/utils/types'
-import { useDispatch, useSelector } from 'react-redux'
-import { setNotReadyToPay } from '@/features/cart/cartSlice'
+import { CartItemProps } from '@/utils/types'
 import PlaceAutocomplete from './AutoCompleteElement'
+import TransbankForm from './TransbankForm'
 
 interface PlaceResult { }
 
@@ -12,12 +11,10 @@ interface PaymentFormProps {
   totalWithVat: number
   children: ReactNode
   setSelectedPlace: (place: PlaceResult | null) => void
+  vat: number
 }
 
-export const PaymentForm = ({ cart, totalWithVat, children, setSelectedPlace }: PaymentFormProps) => {
-
-  const transbank = useSelector((state: StoreProps) => state.cart.transbank)
-  const dispatch = useDispatch()
+export const PaymentForm = ({ cart, vat, totalWithVat, children, setSelectedPlace }: PaymentFormProps) => {
 
   return (
     <main className="w-full flex justify-between">
@@ -26,11 +23,6 @@ export const PaymentForm = ({ cart, totalWithVat, children, setSelectedPlace }: 
           {
             children
           }
-          {
-            cart.map((product, i) => (
-              <SummaryCard product={product} key={i} />
-            ))
-          }
         </div>
         <div className="autocomplete-control">
           <PlaceAutocomplete onPlaceSelect={setSelectedPlace} />
@@ -38,23 +30,25 @@ export const PaymentForm = ({ cart, totalWithVat, children, setSelectedPlace }: 
       </section>
       <section className='flex flex-col gap-y-4 w-[400px] rounded-[20px] min-h-[400px] pt-0'>
         <div className="h-[25px]"></div>
-        <div className='bg-[#ffffff]'>
+        <div className='bg-[#ffffff] mb-5'>
           <img className='w-full' src="https://res.cloudinary.com/maulight/image/upload/v1734712129/zds7cbfpfhfki1djh3wp.png" alt="webpay" />
         </div>
-        <div className="border-b"></div>
-        <div className="flex gap-x-5 mt-10">
-          <h1 className='text-[#ffffff] aktivLight text-[38px] uppercase'>Total</h1>
-          <h1 className='text-[#ffffff] aktiv text-[38px] uppercase'>{totalWithVat}$</h1>
+        {
+          cart.map((product, i) => (
+            <SummaryCard product={product} key={i} />
+          ))
+        }
+        <div>
+          <div className="w-full flex justify-between mt-10">
+            <h1 className='text-[#ffffff] text-[2rem] uppercase'>Total</h1>
+            <h1 className='text-[#ffffff] text-[2rem] uppercase'>{totalWithVat}$</h1>
+          </div>
+          <div className="w-full flex justify-end gap-x-2">
+            <h1 className='text-[#ffffff] text-[1rem] uppercase'>{'(*) Includes'}</h1>
+            <h1 className='text-[#ffffff] text-[1rem] uppercase'>{vat}$ VAT</h1>
+          </div>
         </div>
-        <div className="flex flex-col">
-          <form method="post" action={transbank.url}>
-            <input type="hidden" name="token_ws" value={transbank.token} />
-            <button type='submit' className='w-full h-8 flex justify-center items-center bg-[#ffffff] hover:bg-indigo-500 active:bg-[#ffffff] px-2 uppercase text-[#10100e] mt-3 transition-all duration-200'>
-              Pay
-            </button>
-          </form>
-          <button type='button' onClick={() => { dispatch(setNotReadyToPay()) }} className='h-8 hover:bg-red-600 active:bg-transparent px-2 uppercase text-[#ffffff] mt-3 transition-all duration-200 text-[12px] text-right'>Cancel</button>
-        </div>
+        <TransbankForm />
       </section>
     </main>
   )
