+++ src/routes/Checkout.tsx
@@ -1,31 +1,17 @@
-import { useEffect, useState, type ReactElement } from 'react'
+import { useEffect, type ReactElement } from 'react'
 import { Link } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 import { motion } from 'framer-motion'
 
 //* Components
-import { PaymentForm } from '@/components/checkout/PaymentForm'
 import { CheckSummary } from '@/components/checkout/CheckSummary'
 import { CheckoutCard } from '@/components/checkout/CheckoutCard'
 import { XMarkIcon } from '@heroicons/react/20/solid'
-import { APIProvider, Map, AdvancedMarker, MapCameraChangedEvent } from '@vis.gl/react-google-maps'
 
 //* Types
 import { ProductProps, StoreProps } from '@/utils/types'
 import { fadeIn } from '@/utils/functions'
-
-const apiKey = import.meta.env.VITE_GOOGLE_MAPS_API_KEY
-const mapId = import.meta.env.VITE_GOOGLE_MAPS_MAP_ID
-
-// const geocodeAddress = (address: string, callback: (results: google.maps.GeocoderResult[], status: google.maps.GeocoderStatus) => void) => {
-//     const geocoder = new google.maps.Geocoder()
-//     geocoder.geocode({ address }, (results, status) => {
-//         if (results) {
-//             return callback(results, status)
-//         }
-//         return null
-//     })
-// }
+import GoogleMapsAPI from '@/components/checkout/GoogleMapsAPI'
 
 const Checkout = (): ReactElement => {
     const cart = useSelector((state: StoreProps) => state.cart.cart)
@@ -38,51 +24,11 @@ const Checkout = (): ReactElement => {
     const vat = Math.floor(((total / 100) * 19))
     const totalWithVat = total + vat
 
-    //* Google maps state
-    const [zoom, setZoom] = useState<number>(13)
-    const [geocodeResult, setGeocodeResult] = useState<{ lat: number, lng: number } | null>(null)
-    const [selectedPlace, setSelectedPlace] = useState<google.maps.places.PlaceResult | null>(null)
-
-
-    // const handleGeocode = (e: React.KeyboardEvent<HTMLInputElement>): void => {
-    //     if (e.key === 'Enter') {
-    //         geocodeAddress(address, (results: google.maps.GeocoderResult[], status: google.maps.GeocoderStatus): void => {
-    //             if (status === 'OK') {
-    //                 const newGeoLocation = {
-    //                     lat: results[0].geometry.location.lat(),
-    //                     lng: results[0].geometry.location.lng()
-    //                 }
-    //                 setGeocodeResult(newGeoLocation)
-    //                 setZoom(18)
-    //                 console.log(newGeoLocation, 'This is the new geolocation.')
-    //             } else {
-    //                 console.error('Geocode was not successful for the following reason: ' + status);
-    //             }
-    //         });
-    //     }
-    // }
-
-    const handleCameraChanged = (ev: MapCameraChangedEvent) => {
-        setZoom(ev.detail.zoom)
-        console.log('camera changed:', ev.detail.center, 'zoom:', ev.detail.zoom)
-    }
-
     useEffect(() => {
         if (cart.length > 0) {
             localStorage.setItem('marketplace-cart', JSON.stringify(cart))
         }
     }, [])
-
-    useEffect(() => {
-        if (selectedPlace !== null && selectedPlace.geometry !== undefined && selectedPlace.geometry.location !== undefined) {
-            const newGeoLocation = {
-                lat: selectedPlace.geometry.location.lat(),
-                lng: selectedPlace.geometry.location.lng()
-            }
-            setGeocodeResult(newGeoLocation)
-            setZoom(18)
-        }
-    }, [selectedPlace])
 
     return (
         <div className={`min-[500px]:max-[1440px]:px-10 w-full flex justify-center ${readyToPay ? 'bg-[#10100e]' : 'gap-y-10 bg-[#fdfdfd]'}`}>
@@ -109,25 +55,9 @@ const Checkout = (): ReactElement => {
                     }
                 </div>
                 <>
-
                     {
                         readyToPay ? (
-                            <APIProvider onLoad={() => { console.log('Maps loaded.') }} apiKey={apiKey}>
-                                <PaymentForm setSelectedPlace={setSelectedPlace} cart={cart} totalWithVat={totalWithVat}>
-                                    <>
-                                        <Map
-                                            zoom={zoom}
-                                            mapId={mapId}
-                                            defaultZoom={13}
-                                            center={geocodeResult}
-                                            defaultCenter={{ lat: -33.44888970000001, lng: 289.3307345 }}
-                                            onCameraChanged={handleCameraChanged}>
-                                            {geocodeResult && <AdvancedMarker position={geocodeResult} />}
-                                        </Map>
-                                    </>
-
-                                </PaymentForm>
-                            </APIProvider>
+                            <GoogleMapsAPI vat={vat} totalWithVat={totalWithVat} />
                         )
                             :
                             (
